Conflicts:
	activerecord/lib/active_record/relation/calculations.rb
	activerecord/test/cases/calculations_test.rb

+++ actionpack/test/template/url_helper_test.rb
@@ -90,15 +90,33 @@ class UrlHelperTest < ActiveSupport::TestCase
   def test_button_to_with_javascript_confirm
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      button_to("Hello", "http://www.example.com", :data => { :confirm => "Are you sure?" })
     )
+  end
+
+  def test_button_to_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      )
+    end
   end
 
   def test_button_to_with_javascript_disable_with
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", 'data-disable-with' => "Greeting...")
+      button_to("Hello", "http://www.example.com", :data => { :disable_with => "Greeting..." })
     )
+  end
+
+  def test_button_to_with_javascript_deprecated_disable_with
+    assert_deprecated ":disable_with option is deprecated and will be removed from Rails 4.1. Use ':data => { :disable_with => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :disable_with => "Greeting...")
+      )
+    end
   end
 
   def test_button_to_with_remote_and_form_options
@@ -108,8 +126,33 @@ class UrlHelperTest < ActiveSupport::TestCase
   def test_button_to_with_remote_and_javascript_confirm
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
+      button_to("Hello", "http://www.example.com", :remote => true, :data => { :confirm => "Are you sure?" })
     )
+  end
+
+  def test_button_to_with_remote_and_javascript_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
+      )
+    end
+  end
+
+  def test_button_to_with_remote_and_javascript_disable_with
+    assert_dom_equal(
+      "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
+      button_to("Hello", "http://www.example.com", :remote => true, :data => { :disable_with => "Greeting..." })
+    )
+  end
+
+  def test_button_to_with_remote_and_javascript_deprecated_disable_with
+    assert_deprecated ":disable_with option is deprecated and will be removed from Rails 4.1. Use ':data => { :disable_with => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-disable-with=\"Greeting...\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :remote => true, :disable_with => "Greeting...")
+      )
+    end
   end
 
   def test_button_to_with_remote_false
@@ -201,23 +244,44 @@ class UrlHelperTest < ActiveSupport::TestCase
 
   def test_link_tag_with_custom_onclick
     link = link_to("Hello", "http://www.example.com", :onclick => "alert('yay!')")
-    expected = %{<a href="http://www.example.com" onclick="alert('yay!')">Hello</a>}
+    expected = %{<a href="http://www.example.com" onclick="alert(&#x27;yay!&#x27;)">Hello</a>}
     assert_dom_equal expected, link
   end
 
   def test_link_tag_with_javascript_confirm
     assert_dom_equal(
       "<a href=\"http://www.example.com\" data-confirm=\"Are you sure?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "Are you sure?" })
     )
     assert_dom_equal(
-      "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure, can you?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure, can you?")
+      "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure, can you?\">Hello</a>",
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "You cant possibly be sure, can you?" })
     )
     assert_dom_equal(
-      "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure,\n can you?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure,\n can you?")
+      "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure,\n can you?\">Hello</a>",
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "You cant possibly be sure,\n can you?" })
     )
+  end
+
+  def test_link_tag_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"Are you sure?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      )
+    end
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure, can you?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "You cant possibly be sure, can you?")
+      )
+    end
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"You cant possibly be sure,\n can you?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "You cant possibly be sure,\n can you?")
+      )
+    end
   end
 
   def test_link_to_with_remote
@@ -265,16 +329,35 @@ class UrlHelperTest < ActiveSupport::TestCase
   def test_link_tag_using_post_javascript_and_confirm
     assert_dom_equal(
       "<a href=\"http://www.example.com\" data-method=\"post\" rel=\"nofollow\" data-confirm=\"Are you serious?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :method => :post, :confirm => "Are you serious?")
+      link_to("Hello", "http://www.example.com", :method => :post, :data => { :confirm => "Are you serious?" })
     )
+  end
+
+  def test_link_tag_using_post_javascript_and_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-method=\"post\" rel=\"nofollow\" data-confirm=\"Are you serious?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :method => :post, :confirm => "Are you serious?")
+      )
+    end
   end
 
   def test_link_tag_using_delete_javascript_and_href_and_confirm
     assert_dom_equal(
       "<a href='\#' rel=\"nofollow\" data-confirm=\"Are you serious?\" data-method=\"delete\">Destroy</a>",
-      link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :confirm => "Are you serious?"),
+      link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :data => { :confirm => "Are you serious?" }),
       "When specifying url, form should be generated with it, but not this.href"
     )
+  end
+
+  def test_link_tag_using_delete_javascript_and_href_and_with_deprecated_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.1. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href='\#' rel=\"nofollow\" data-confirm=\"Are you serious?\" data-method=\"delete\">Destroy</a>",
+        link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :confirm => "Are you serious?"),
+        "When specifying url, form should be generated with it, but not this.href"
+      )
+    end
   end
 
   def test_link_tag_with_block
